// frontend/src/util/type/Notification.enum.ts
export enum NotificationEumn {
  CREATED_QUESTION = "CREATED_QUESTION",
  UPDATED_QUESTION = "UPDATED_QUESTION",
  DELETED_QUESTION = "DELETED_QUESTION",

  CREATED_ANSWER = "CREATED_ANSWER",
  UPDATED_ANSWER = "UPDATED_ANSWER",
  DELETED_ANSWER = "DELETED_ANSWER",
  ACCEPTED_ANSWER = "ACCEPTED_ANSWER",
  UN_APPROVED_ANSWER = "UN_APPROVED_ANSWER",

  CREATED_COMMENT = "CREATED_COMMENT",
  UPDATED_COMMENT = "UPDATED_COMMENT",
  DELETED_COMMENT = "DELETED_COMMENT",

  UP_VOTE = "UP_VOTE",
  DOWN_VOTE = "DOWN_VOTE",
  CANCEL_UP_VOTE = "CANCEL_UP_VOTE",
  CANCEL_DOWN_VOTE = "CANCEL_DOWN_VOTE",
  CHANGE_VOTE_UP_TO_DOWN = "CHANGE_VOTE_UP_TO_DOWN",
  CHANGE_VOTE_DOWN_TO_UP = "CHANGE_VOTE_DOWN_TO_UP",
}
export enum NotificationDesEumn {
  DESCRIPTION_CREATED_QUESTION = "DESCRIPTION_CREATED_QUESTION",
  DESCRIPTION_UPDATED_QUESTION = "DESCRIPTION_UPDATED_QUESTION",
  DESCRIPTION_DELETED_QUESTION = "DESCRIPTION_DELETED_QUESTION",

  DESCRIPTION_CREATED_ANSWER = "DESCRIPTION_CREATED_ANSWER",
  DESCRIPTION_UPDATED_ANSWER = "DESCRIPTION_UPDATED_ANSWER",
  DESCRIPTION_DELETED_ANSWER = "DESCRIPTION_DELETED_ANSWER",
  DESCRIPTION_ACCEPTED_ANSWER = "DESCRIPTION_ACCEPTED_ANSWER",
  DESCRIPTION_UN_APPROVED_ANSWER = "DESCRIPTION_UN_APPROVED_ANSWER",

  DESCRIPTION_CREATED_COMMENT = "DESCRIPTION_CREATED_COMMENT",
  DESCRIPTION_UPDATED_COMMENT = "DESCRIPTION_UPDATED_COMMENT",
  DESCRIPTION_DELETED_COMMENT = "DESCRIPTION_DELETED_COMMENT",

  DESCRIPTION_UP_VOTE = "DESCRIPTION_UP_VOTE",
  DESCRIPTION_DOWN_VOTE = "DESCRIPTION_DOWN_VOTE",
  DESCRIPTION_CANCEL_UP_VOTE = "DESCRIPTION_CANCEL_UP_VOTE",
  DESCRIPTION_CANCEL_DOWN_VOTE = "DESCRIPTION_CANCEL_DOWN_VOTE",
  DESCRIPTION_CHANGE_VOTE_UP_TO_DOWN = "DESCRIPTION_CHANGE_VOTE_UP_TO_DOWN",
  DESCRIPTION_CHANGE_VOTE_DOWN_TO_UP_VOTE = "DESCRIPTION_CHANGE_VOTE_DOWN_TO_UP_VOTE",
}